Abstract TreeNode and TreeNodeFixedSize
