+++ community/index/src/main/java/org/neo4j/index/internal/gbptree/GBPTree.java
@@ -425,7 +425,7 @@ public class GBPTree<KEY,VALUE> implements Closeable
             this.pagedFile = openOrCreate( pageCache, indexFile, tentativePageSize, layout );
             this.pageSize = pagedFile.pageSize();
             closed = false;
-            this.bTreeNode = new TreeNode<>( pageSize, layout );
+            this.bTreeNode = new TreeNodeFixedSize<>( pageSize, layout );
             this.freeList = new FreeListIdProvider( pagedFile, pageSize, rootId, FreeListIdProvider.NO_MONITOR );
             this.writer = new SingleWriter( new InternalTreeLogic<>( freeList, bTreeNode, layout ) );
