+++ community/index/src/main/java/org/neo4j/index/internal/gbptree/TreeNode.java
@@ -26,7 +26,6 @@ import org.neo4j.io.pagecache.PageCursor;
 
 import static org.neo4j.index.internal.gbptree.GenerationSafePointerPair.NO_LOGICAL_POS;
 import static org.neo4j.index.internal.gbptree.GenerationSafePointerPair.read;
-import static org.neo4j.index.internal.gbptree.TreeNode.Type.INTERNAL;
 
 /**
  * Methods to manipulate single tree node such as set and get header fields,
@@ -65,7 +64,7 @@ import static org.neo4j.index.internal.gbptree.TreeNode.Type.INTERNAL;
  * @param <KEY> type of key
  * @param <VALUE> type of value
  */
-class TreeNode<KEY,VALUE>
+abstract class TreeNode<KEY,VALUE>
 {
     enum Type
     {
@@ -87,39 +86,17 @@ class TreeNode<KEY,VALUE>
     static final int BYTE_POS_SUCCESSOR = BYTE_POS_LEFTSIBLING + SIZE_PAGE_REFERENCE;
     static final int HEADER_LENGTH = BYTE_POS_SUCCESSOR + SIZE_PAGE_REFERENCE;
 
-    private static final byte LEAF_FLAG = 1;
+    static final byte LEAF_FLAG = 1;
     static final byte INTERNAL_FLAG = 0;
     static final long NO_NODE_FLAG = 0;
 
+    final Layout<KEY,VALUE> layout;
     private final int pageSize;
-    private final int internalMaxKeyCount;
-    private final int leafMaxKeyCount;
-    private final Layout<KEY,VALUE> layout;
-
-    private final int keySize;
-    private final int valueSize;
 
     TreeNode( int pageSize, Layout<KEY,VALUE> layout )
     {
         this.pageSize = pageSize;
         this.layout = layout;
-        this.keySize = layout.keySize();
-        this.valueSize = layout.valueSize();
-        this.internalMaxKeyCount = Math.floorDiv( pageSize - (HEADER_LENGTH + SIZE_PAGE_REFERENCE),
-                keySize + SIZE_PAGE_REFERENCE);
-        this.leafMaxKeyCount = Math.floorDiv( pageSize - HEADER_LENGTH, keySize + valueSize );
-
-        if ( internalMaxKeyCount < 2 )
-        {
-            throw new MetadataMismatchException(
-                    "For layout %s a page size of %d would only fit %d internal keys, minimum is 2",
-                    layout, pageSize, internalMaxKeyCount );
-        }
-        if ( leafMaxKeyCount < 2 )
-        {
-            throw new MetadataMismatchException( "A page size of %d would only fit leaf keys, minimum is 2",
-                    pageSize, leafMaxKeyCount );
-        }
     }
 
     static byte nodeType( PageCursor cursor )
@@ -241,197 +218,53 @@ class TreeNode<KEY,VALUE>
 
     // BODY METHODS
 
-    KEY keyAt( PageCursor cursor, KEY into, int pos, Type type )
-    {
-        cursor.setOffset( keyOffset( pos ) );
-        layout.readKey( cursor, into );
-        return into;
-    }
+    abstract KEY keyAt( PageCursor cursor, KEY into, int pos, Type type );
 
     // Insert key without associated value.
     // Useful for internal nodes and testing.
-    void insertKeyAt( PageCursor cursor, KEY key, int pos, int keyCount, Type type )
-    {
-        insertKeySlotsAt( cursor, pos, 1, keyCount );
-        cursor.setOffset( keyOffset( pos ) );
-        layout.writeKey( cursor, key );
-    }
+    abstract void insertKeyAt( PageCursor cursor, KEY key, int pos, int keyCount, Type type );
 
-    void insertKeyValueAt( PageCursor cursor, KEY key, VALUE value, int pos, int keyCount )
-    {
-        insertKeyAt( cursor, key, pos, keyCount, Type.LEAF );
-        insertValueAt( cursor, value, pos, keyCount );
-    }
+    abstract void insertKeyValueAt( PageCursor cursor, KEY key, VALUE value, int pos, int keyCount );
 
     // Remove key without removing associated value.
     // Useful for internal nodes and testing.
-    void removeKeyAt( PageCursor cursor, int pos, int keyCount, Type type )
-    {
-        removeSlotAt( cursor, pos, keyCount, keyOffset( 0 ), keySize );
-    }
+    abstract void removeKeyAt( PageCursor cursor, int pos, int keyCount, Type type );
 
-    void removeKeyValueAt( PageCursor cursor, int pos, int keyCount )
-    {
-        removeKeyAt( cursor, pos, keyCount, Type.LEAF );
-        removeValueAt( cursor, pos, keyCount );
-    }
+    abstract void removeKeyValueAt( PageCursor cursor, int pos, int keyCount );
 
-    void setKeyAt( PageCursor cursor, KEY key, int pos, Type type )
-    {
-        cursor.setOffset( keyOffset( pos ) );
-        layout.writeKey( cursor, key );
-    }
+    abstract void setKeyAt( PageCursor cursor, KEY key, int pos, Type type );
 
-    VALUE valueAt( PageCursor cursor, VALUE value, int pos )
-    {
-        cursor.setOffset( valueOffset( pos ) );
-        layout.readValue( cursor, value );
-        return value;
-    }
+    abstract VALUE valueAt( PageCursor cursor, VALUE value, int pos );
 
-    void setValueAt( PageCursor cursor, VALUE value, int pos )
-    {
-        cursor.setOffset( valueOffset( pos ) );
-        layout.writeValue( cursor, value );
-    }
+    abstract void setValueAt( PageCursor cursor, VALUE value, int pos );
 
-    long childAt( PageCursor cursor, int pos, long stableGeneration, long unstableGeneration )
-    {
-        cursor.setOffset( childOffset( pos ) );
-        return read( cursor, stableGeneration, unstableGeneration, pos );
-    }
+    abstract long childAt( PageCursor cursor, int pos, long stableGeneration, long unstableGeneration );
 
-    void insertChildAt( PageCursor cursor, long child, int pos, int keyCount,
-            long stableGeneration, long unstableGeneration )
-    {
-        insertChildSlotsAt( cursor, pos, 1, keyCount );
-        setChildAt( cursor, child, pos, stableGeneration, unstableGeneration );
-    }
+    abstract void insertChildAt( PageCursor cursor, long child, int pos, int keyCount,
+            long stableGeneration, long unstableGeneration );
 
-    void removeChildAt( PageCursor cursor, int pos, int keyCount )
-    {
-        removeSlotAt( cursor, pos, keyCount + 1, childOffset( 0 ), childSize() );
-    }
+    abstract void removeChildAt( PageCursor cursor, int pos, int keyCount );
 
-    void setChildAt( PageCursor cursor, long child, int pos, long stableGeneration, long unstableGeneration )
-    {
-        cursor.setOffset( childOffset( pos ) );
-        writeChild( cursor, child, stableGeneration, unstableGeneration );
-    }
+    abstract void setChildAt( PageCursor cursor, long child, int pos, long stableGeneration, long unstableGeneration );
 
-    private static void writeChild( PageCursor cursor, long child, long stableGeneration, long unstableGeneration )
+    static void writeChild( PageCursor cursor, long child, long stableGeneration, long unstableGeneration )
     {
         GenerationSafePointerPair.write( cursor, child, stableGeneration, unstableGeneration );
     }
 
-    private void insertKeyValueSlotsAt( PageCursor cursor, int pos, int numberOfSlots, int keyCount )
-    {
-        insertKeySlotsAt( cursor, pos, numberOfSlots, keyCount );
-        insertValueSlotsAt( cursor, pos, numberOfSlots, keyCount );
-    }
+    abstract int internalMaxKeyCount();
 
-    // Always insert together with key. Use insertKeyValueAt
-    private void insertValueAt( PageCursor cursor, VALUE value, int pos, int keyCount )
-    {
-        insertValueSlotsAt( cursor, pos, 1, keyCount );
-        setValueAt( cursor, value, pos );
-    }
-
-    // Always insert together with key. Use removeKeyValueAt
-    private void removeValueAt( PageCursor cursor, int pos, int keyCount )
-    {
-        removeSlotAt( cursor, pos, keyCount, valueOffset( 0 ), valueSize );
-    }
-
-    /**
-     * Moves items (key/value/child) one step to the right, which means rewriting all items of the particular type
-     * from pos - itemCount.
-     * itemCount is keyCount for key and value, but keyCount+1 for children.
-     */
-    private static void insertSlotsAt( PageCursor cursor, int pos, int numberOfSlots, int itemCount, int baseOffset,
-            int itemSize )
-    {
-        for ( int posToMoveRight = itemCount - 1, offset = baseOffset + posToMoveRight * itemSize;
-              posToMoveRight >= pos; posToMoveRight--, offset -= itemSize )
-        {
-            cursor.copyTo( offset, cursor, offset + itemSize * numberOfSlots, itemSize );
-        }
-    }
-
-    private static void removeSlotAt( PageCursor cursor, int pos, int itemCount, int baseOffset, int itemSize )
-    {
-        for ( int posToMoveLeft = pos + 1, offset = baseOffset + posToMoveLeft * itemSize;
-              posToMoveLeft < itemCount; posToMoveLeft++, offset += itemSize )
-        {
-            cursor.copyTo( offset, cursor, offset - itemSize, itemSize );
-        }
-    }
-
-    private void insertKeySlotsAt( PageCursor cursor, int pos, int numberOfSlots, int keyCount )
-    {
-        insertSlotsAt( cursor, pos, numberOfSlots, keyCount, keyOffset( 0 ), keySize );
-    }
-
-    private void insertValueSlotsAt( PageCursor cursor, int pos, int numberOfSlots, int keyCount )
-    {
-        insertSlotsAt( cursor, pos, numberOfSlots, keyCount, valueOffset( 0 ), valueSize );
-    }
-
-    private void insertChildSlotsAt( PageCursor cursor, int pos, int numberOfSlots, int keyCount )
-    {
-        insertSlotsAt( cursor, pos, numberOfSlots, keyCount + 1, childOffset( 0 ), childSize() );
-    }
-
-    int internalMaxKeyCount()
-    {
-        return internalMaxKeyCount;
-    }
-
-    int leafMaxKeyCount()
-    {
-        return leafMaxKeyCount;
-    }
+    abstract int leafMaxKeyCount();
 
     // HELPERS
 
-    boolean reasonableKeyCount( int keyCount )
-    {
-        return keyCount >= 0 && keyCount <= Math.max( internalMaxKeyCount(), leafMaxKeyCount() );
-    }
+    abstract boolean reasonableKeyCount( int keyCount );
 
-    private int keyOffset( int pos )
-    {
-        return HEADER_LENGTH + pos * keySize;
-    }
-
-    private int valueOffset( int pos )
-    {
-        return HEADER_LENGTH + leafMaxKeyCount * keySize + pos * valueSize;
-    }
-
-    int childOffset( int pos )
-    {
-        return HEADER_LENGTH + internalMaxKeyCount * keySize + pos * SIZE_PAGE_REFERENCE;
-    }
+    abstract int childOffset( int pos );
 
     static boolean isNode( long node )
     {
         return GenerationSafePointerPair.pointer( node ) != NO_NODE_FLAG;
-    }
-
-    private int keySize()
-    {
-        return keySize;
-    }
-
-    private int valueSize()
-    {
-        return valueSize;
-    }
-
-    private static int childSize()
-    {
-        return SIZE_PAGE_REFERENCE;
     }
 
     Comparator<KEY> keyComparator()
@@ -448,36 +281,23 @@ class TreeNode<KEY,VALUE>
     @Override
     public String toString()
     {
-        return "TreeNode[pageSize:" + pageSize + ", internalMax:" + internalMaxKeyCount +
-                ", leafMax:" + leafMaxKeyCount + ", keySize:" + keySize + ", valueSize:" + valueSize + "]";
+        return "TreeNode[pageSize:" + pageSize + ", internalMax:" + internalMaxKeyCount() + ", leafMax:" + leafMaxKeyCount() + ", " +
+                additionalToString() + "]";
     }
+
+    abstract String additionalToString();
 
     /* SPLIT, MERGE AND REBALANCE */
 
-    boolean internalOverflow( int keyCount )
-    {
-        return keyCount > internalMaxKeyCount();
-    }
+    abstract boolean internalOverflow( int keyCount );
 
-    boolean leafOverflow( int keyCount )
-    {
-        return keyCount > leafMaxKeyCount();
-    }
+    abstract boolean leafOverflow( int keyCount );
 
-    boolean leafUnderflow( int keyCount )
-    {
-        return keyCount < (leafMaxKeyCount() + 1) / 2;
-    }
+    abstract boolean leafUnderflow( int keyCount );
 
-    boolean canRebalanceLeaves( int leftKeyCount, int rightKeyCount )
-    {
-        return leftKeyCount + rightKeyCount >= leafMaxKeyCount();
-    }
+    abstract boolean canRebalanceLeaves( int leftKeyCount, int rightKeyCount );
 
-    boolean canMergeLeaves( int leftKeyCount, int rightKeyCount )
-    {
-        return leftKeyCount + rightKeyCount <= leafMaxKeyCount();
-    }
+    abstract boolean canMergeLeaves( int leftKeyCount, int rightKeyCount );
 
     /**
      * Performs the entry moving part of split in leaf.
@@ -486,42 +306,9 @@ class TreeNode<KEY,VALUE>
      *
      * Key count is updated.
      */
-    void doSplitLeaf( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount, int insertPos, KEY newKey,
-            VALUE newValue, int middlePos )
-    {
-        if ( insertPos < middlePos )
-        {
-            //                  v-------v       copy
-            // before _,_,_,_,_,_,_,_,_,_
-            // insert _,_,_,X,_,_,_,_,_,_,_
-            // middle           ^
-            copyKeysAndValues( leftCursor, middlePos - 1, rightCursor, 0, rightKeyCount );
-            insertKeyValueAt( leftCursor, newKey, newValue, insertPos, middlePos - 1 );
-        }
-        else
-        {
-            //                  v---v           first copy
-            //                        v-v       second copy
-            // before _,_,_,_,_,_,_,_,_,_
-            // insert _,_,_,_,_,_,_,_,X,_,_
-            // middle           ^
-            int countBeforePos = insertPos - middlePos;
-            if ( countBeforePos > 0 )
-            {
-                // first copy
-                copyKeysAndValues( leftCursor, middlePos, rightCursor, 0, countBeforePos );
-            }
-            insertKeyValueAt( rightCursor, newKey, newValue, countBeforePos, countBeforePos );
-            int countAfterPos = leftKeyCount - insertPos;
-            if ( countAfterPos > 0 )
-            {
-                // second copy
-                copyKeysAndValues( leftCursor, insertPos, rightCursor, countBeforePos + 1, countAfterPos );
-            }
-        }
-        TreeNode.setKeyCount( leftCursor, middlePos );
-        TreeNode.setKeyCount( rightCursor, rightKeyCount );
-    }
+    abstract void doSplitLeaf( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount, int insertPos,
+            KEY newKey,
+            VALUE newValue, int middlePos );
 
     /**
      * Performs the entry moving part of split in internal.
@@ -530,103 +317,16 @@ class TreeNode<KEY,VALUE>
      *
      * Key count is updated.
      */
-    void doSplitInternal( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount, int insertPos, KEY newKey,
-            long newRightChild, int middlePos, long stableGeneration, long unstableGeneration )
-    {
-        if ( insertPos < middlePos )
-        {
-            //                         v-------v       copy
-            // before key    _,_,_,_,_,_,_,_,_,_
-            // before child -,-,-,-,-,-,-,-,-,-,-
-            // insert key    _,_,X,_,_,_,_,_,_,_,_
-            // insert child -,-,-,x,-,-,-,-,-,-,-,-
-            // middle key              ^
-
-            leftCursor.copyTo( keyOffset( middlePos ), rightCursor, keyOffset( 0 ), rightKeyCount * keySize() );
-            leftCursor.copyTo( childOffset( middlePos ), rightCursor, childOffset( 0 ), (rightKeyCount + 1) * childSize() );
-            insertKeyAt( leftCursor, newKey, insertPos, middlePos - 1, INTERNAL );
-            insertChildAt( leftCursor, newRightChild, insertPos + 1, middlePos - 1, stableGeneration, unstableGeneration );
-        }
-        else
-        {
-            // pos > middlePos
-            //                         v-v          first copy
-            //                             v-v-v    second copy
-            // before key    _,_,_,_,_,_,_,_,_,_
-            // before child -,-,-,-,-,-,-,-,-,-,-
-            // insert key    _,_,_,_,_,_,_,X,_,_,_
-            // insert child -,-,-,-,-,-,-,-,x,-,-,-
-            // middle key              ^
-
-            // pos == middlePos
-            //                                      first copy
-            //                         v-v-v-v-v    second copy
-            // before key    _,_,_,_,_,_,_,_,_,_
-            // before child -,-,-,-,-,-,-,-,-,-,-
-            // insert key    _,_,_,_,_,X,_,_,_,_,_
-            // insert child -,-,-,-,-,-,x,-,-,-,-,-
-            // middle key              ^
-
-            // Keys
-            int countBeforePos = insertPos - (middlePos + 1);
-            // ... first copy
-            if ( countBeforePos > 0 )
-            {
-                leftCursor.copyTo( keyOffset( middlePos + 1 ), rightCursor, keyOffset( 0 ), countBeforePos * keySize() );
-            }
-            // ... insert
-            if ( countBeforePos >= 0 )
-            {
-                insertKeyAt( rightCursor, newKey, countBeforePos, countBeforePos, INTERNAL );
-            }
-            // ... second copy
-            int countAfterPos = leftKeyCount - insertPos;
-            if ( countAfterPos > 0 )
-            {
-                leftCursor.copyTo( keyOffset( insertPos ), rightCursor, keyOffset( countBeforePos + 1 ), countAfterPos * keySize() );
-            }
-
-            // Children
-            countBeforePos = insertPos - middlePos;
-            // ... first copy
-            if ( countBeforePos > 0 )
-            {
-                // first copy
-                leftCursor.copyTo( childOffset( middlePos + 1 ), rightCursor, childOffset( 0 ), countBeforePos * childSize() );
-            }
-            // ... insert
-            insertChildAt( rightCursor, newRightChild, countBeforePos, countBeforePos, stableGeneration, unstableGeneration );
-            // ... second copy
-            if ( countAfterPos > 0 )
-            {
-                leftCursor.copyTo( childOffset( insertPos + 1 ), rightCursor, childOffset( countBeforePos + 1 ),
-                        countAfterPos * childSize() );
-            }
-        }
-        TreeNode.setKeyCount( leftCursor, middlePos );
-        TreeNode.setKeyCount( rightCursor, rightKeyCount );
-    }
+    abstract void doSplitInternal( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount, int insertPos,
+            KEY newKey,
+            long newRightChild, int middlePos, long stableGeneration, long unstableGeneration );
 
     /**
      * Move all rightmost keys and values in left node from given position to right node.
      *
      * Key count is NOT updated.
      */
-    void moveKeyValuesFromLeftToRight( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount,
-            int fromPosInLeftNode )
-    {
-        int numberOfKeysToMove = leftKeyCount - fromPosInLeftNode;
+    abstract void moveKeyValuesFromLeftToRight( PageCursor leftCursor, int leftKeyCount, PageCursor rightCursor, int rightKeyCount,
+            int fromPosInLeftNode );
 
-        // Push keys and values in right sibling to the right
-        insertKeyValueSlotsAt( rightCursor, 0, numberOfKeysToMove, rightKeyCount );
-
-        // Move keys and values from left sibling to right sibling
-        copyKeysAndValues( leftCursor, fromPosInLeftNode, rightCursor, 0, numberOfKeysToMove );
-    }
-
-    private void copyKeysAndValues( PageCursor fromCursor, int fromPos, PageCursor toCursor, int toPos, int count )
-    {
-        fromCursor.copyTo( keyOffset( fromPos ), toCursor, keyOffset( toPos ), count * keySize() );
-        fromCursor.copyTo( valueOffset( fromPos ), toCursor, valueOffset( toPos ),count * valueSize() );
-    }
 }
